pyproject.toml
[project]
name = "api-manager-orders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "fastapi[standard]>=0.120.3",
    "greenlet>=3.2.4",
    "jose>=1.0.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
]
